Warning removed.
git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.6.0.stable@234 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Utils/Logger.cpp
@@ -75,7 +75,7 @@ void StdLogger::doLog(level_t l,const std::string & msg){
 		case LOG_ALL:
 		case LOG_NONE:
 		default:{
-			out << "Logging ("<<l<<"):";
+			out << "Logging ("<<static_cast<int>(l)<<"):";
 		}
 	}
 	out << msg << std::endl;
